Define addition operators on Content for concatenating strings and other Content; move Content from Element to Fundament; define .Content() extension method on RenderFragment to convert a RenderFragment to Content; make Button Bits inline-flex and align items to center; change BitPixels Bit getter to return a double to support sub-pixel sizes

+++ Element/BitGetters.cs
@@ -30,7 +30,7 @@ namespace Integrant.Element
 
         public delegate double BitREM();
 
-        public delegate uint BitPixels();
+        public delegate double BitPixels();
 
         public delegate ushort BitWeight();
